MongoDb Access
Local storage is no longer in use. instead of local storage python was added so the program can handle MongoDb as it's database. the structure of the program needed to change so the flask could be run without any problem. also Javascript was updated with the details inside the routes of the flask.